Fix conflict with Jisho open sentences hint mode
The hint mode for "open sentence examples in new tab" (;G) conflicts
with that for SVG-pan-zoom (although the latter is still just a
placeholder).

Since the available extended hint modes are limited, it's wasteful to
have separate hint modes for "open sentence examples" and "open
sentence examples in new tab". If the former is defined, the latter
can be achieved by prefixing with <C-t>.

Thus, remove the modes ;g and ;G for sentence examples and replace
with ;e.

Unfortunately, attempts at bindings like:
    map -group jisho <site>gx -ex tab normal ;g
don't work.