Remove <Space><Space> binding to prevent conflicts
Binding <Space><Space> slows down group-specific bindings of <Space>,
since the global binding is still in effect for the group.

I don't know how to unmap a binding on a group-specific basis.
Attempting to unmap <Space><Space> in the group results in the expected
"No such mapping" error.

Therefore, it's necessary to avoid global bindings if a prefix of the
binding is used as a group-specific binding.

Replace the <Space><Space> binding for autopager with <C-Space>.