Coursera: add hint mode & binding to jump to matches
Add non-ascii hint-mode ẖ to match elements with class "highlighted"
and click them when chosen. This hint mode allows selection of search
matches in the interactive transcript, and jumps to the selected text
when a hint is chosen.

Bind <site>f to activate this hint mode.

A note is added to the hints section listing this mode.

Additionally, add headings and spacing in group coursera.