Coursera: add binding to scroll next match into view
Bind <site><Tab> to call courseraScrollToNextHighlighted().
This scrolls the next match in the interactive transcript into view
at the bottom of the page. If no further matches remain, it scrolls
back to the first match.

The alignToTop parameter of scrollIntoView() is set to false to
avoid the text being obscured by the navbar at the top.

Note: the window height is retrieved using content.window.innerHeight
It is common to obtain the viewport height with:
    window.innerHeight || document.documentElement.clientHeight
However window.innerHeight alone seems sufficient here.