Add functions `listElements`, `listNodes` and `matches`
`listElements` and `listNodes` return an array of elements/nodes which
satisfy a given predicate. Optionally, they may be restricted to a subtree
under a top element/node, and matches may be restricted to a maximum number.

`matches` is a convenience function for creating predicates. It checks that
the argument exists, has a `match` method, and matches a given regex/string.

Add functions to `window` for command-line use.