Move hints to just before site-specific code
Functions passed to `hints.addMode` must be defined before the statement is
executed. Moving the hints section to after the function definitions removes
the need for indirection such as e=>myFunc(e).

Moving hints section after the site-specific code breaks the hinting for
`about:newtab`, so move it just before that section.